+++ website/projects/scimaps/src/app/app.module.ts
@@ -5,7 +5,7 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { CoreModule } from './core/core.module';
-import { NewsItemModule } from './shared/news-item/news-item.module';
+import { NewsItemModule } from './shared/components/news-item/news-item.module';
 
 
 @NgModule({
