+++ website/projects/scimaps/src/app/shared/components/news-item/news-item.component.html
@@ -1,0 +1,8 @@
+<a class="news-item" [href]="item.pdfLink">
+  <img class="thumb" [src]="item.thumbnail" alt="image thumbnail">
+  <div class="info">
+    <div class="date">{{item.date | date}}</div>
+    <div class="title">{{item.title}}</div>
+    <div class="identifier">{{item.source}} at {{item.institution}}</div>
+  </div>
+</a>
