+++ website/projects/scimaps/src/app/shared/components/news-item/news-item.component.scss
@@ -1,0 +1,43 @@
+:host {
+  display: block;
+
+  .news-item {
+    display: flex;
+    box-shadow: 0 1px 3px rgba(0, 0, 0, 0.54);
+    text-decoration: none;
+    margin-top: 0.5rem;
+    margin-bottom: 0.5rem;
+    border-radius: 0.125rem;
+
+    .thumb {
+      width: 9rem;
+      object-fit: cover;
+      border-width: 2px;
+      border-style: solid;
+    }
+
+    .date {
+      margin: 0;
+      text-transform: uppercase;
+    }
+
+    .info {
+      display: flex;
+      flex-direction: column;
+      padding: 1rem;
+
+      .title {
+        margin-top: 0.5rem;
+        margin-bottom: 0.5rem;
+      }
+
+      .identifier {
+        margin: 0;
+      }
+    }
+  }
+
+  .news-item:hover {
+    box-shadow: 0 2px 6px rgba(0, 0, 0, 0.54);
+  }
+}
