+++ website/projects/scimaps/src/app/shared/components/news-item/news-item.module.ts
@@ -1,5 +1,5 @@
-import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
 
 import { NewsItemComponent } from './news-item.component';
