--- website/projects/scimaps/src/app/shared/news-item/news-item.component.html
@@ -1,8 +1,0 @@
-<a class="news-item" [href]="item.pdfLink">
-  <img class="thumb" [src]="item.thumbnail" alt="image thumbnail">
-  <div class="info">
-    <div class="mat-display-1 date">{{item.date}}</div>
-    <div class="mat-subheading-1 title">{{item.title}}</div>
-    <div class="mat-subheading-2 identifier">{{item.source}} at {{item.institution}}</div>
-  </div>
-</a>
