+++ website/projects/scimaps/src/app/shared/shared.theme.scss
@@ -1,5 +1,5 @@
-@use './news-item/news-item.theme.scss' as newsItem;
+@use './components/news-item/news-item.theme.scss';
 
 @mixin theme($theme-or-config) {
-  @include newsItem.theme($theme-or-config);
+  @include news-item.theme($theme-or-config);
 }
